checked in production code for Token, Staking & Sale
as part of the final check to make the code public we have updated folder structures and file locations were changed around to make them more meaningful and follow some industry best practices. We have checked and updated imports where needed but there may be a few left behind.

also updated the README.md to be in alignment with what was deployed and the current state of the project